support entries for storage map
